refactor(scheduler): move GatherSchedulerLogs to its own file (#701)
